update changelog and move GmawProcess classes, reorder inits
