+++ CHANGELOG.md
@@ -5,6 +5,21 @@
 - Added `weldx.utility.ureg_check_class` class decorator to enable `pint` dimensionality checks with `@dataclass`. [[#179]](https://github.com/BAMWelDX/weldx/pull/179)
 - Made coordinates and orientations optional for LCS schema. Missing values are interpreted as unity translation/rotation. An empty LCS object represents a unity transformation step. [[#177]](https://github.com/BAMWelDX/weldx/pull/177)
 
+### changes
+- refactor welding groove classes [[#181]](https://github.com/BAMWelDX/weldx/pull/181)
+  - refactor groove codebase to make use of subclasses and classnames for more generic functions
+  - add `_meta` attribute to subclasses that map class attributes (dataclass parameters) to common names
+  - rework `get_groove` to make use of new class layout and parse function arguments
+- create `weldx.welding` module (contains GMAW processes and groove definitions) [[#181]](https://github.com/BAMWelDX/weldx/pull/181)
+- move `GmawProcessTypeAsdf` to `asdf.tags` folder [[#181]](https://github.com/BAMWelDX/weldx/pull/181)
+- reorder module imports in `weldx.__init__` [[#181]](https://github.com/BAMWelDX/weldx/pull/181)
+
+### ASDF
+- refactor ISO 9692-1 groove schema definitions and classes [[#181]](https://github.com/BAMWelDX/weldx/pull/181)
+  - move base schema definitions in file `terms-1.0.0.yaml` to `weldx/groove`
+  - split old schema into multiple files (1 per groove type) and create folder `iso_9692_1_2013_12`
+
+
 ## 0.2.1 (26.10.2020)
 ### changes
 - Documentation
