+++ weldx/__init__.py
@@ -1,17 +1,19 @@
 """
 isort:skip_file
 """
-
 import warnings
 
-import weldx.utility
-import weldx.welding
-
-# main packages
+# main modules
+import weldx.utility  # import this first to avoid circular dependencies
 import weldx.core
 import weldx.geometry
 import weldx.transformations
+import weldx.welding
 
+# asdf extensions and tags
+import weldx.asdf
+
+# class imports to weldx namespace
 from weldx.constants import WELDX_QUANTITY as Q_
 from weldx.core import MathematicalExpression, TimeSeries
 from weldx.geometry import ArcSegment, Geometry, LineSegment, Profile, Shape, Trace
@@ -20,12 +22,7 @@ from weldx.transformations import (
     LocalCoordinateSystem,
     WXRotation,
 )
-
-# class imports to weldx namespace
 from weldx.welding.groove.iso_9692_1 import get_groove
-
-# asdf extensions and tags
-import weldx.asdf
 
 # versioneer
 from ._version import get_versions
