+++ weldx/welding/__init__.py
@@ -1,4 +1,3 @@
 """Temporary module for welding related classes."""
 
 from . import groove
-from .processes import GmawProcess
